# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Audit Web IA API",
    "version": "0.1.0"
  },
  "paths": {
    "/audit": {
      "post": {
        "tags": [
          "audit"
        ],
        "summary": "Run Audit",
        "description": "Create a pending job and launch the pipeline in the background.",
        "operationId": "run_audit_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditJobSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/competitor": {
      "get": {
        "tags": [
          "competitor"
        ],
        "summary": "List Battles",
        "operationId": "list_battles_competitor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompetitorBattle"
                  },
                  "type": "array",
                  "title": "Response List Battles Competitor Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "competitor"
        ],
        "summary": "Start Battle",
        "operationId": "start_battle_competitor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitorBattleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitorBattle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/content-brief": {
      "get": {
        "tags": [
          "content-brief"
        ],
        "summary": "List Briefs",
        "operationId": "list_briefs_content_brief_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContentBrief"
                  },
                  "type": "array",
                  "title": "Response List Briefs Content Brief Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "content-brief"
        ],
        "summary": "Start Brief",
        "operationId": "start_brief_content_brief_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentBriefRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentBrief"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/ai-visibility": {
      "get": {
        "tags": [
          "ai-visibility"
        ],
        "summary": "List Checks",
        "operationId": "list_checks_ai_visibility_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AiVisibilityCheck"
                  },
                  "type": "array",
                  "title": "Response List Checks Ai Visibility Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "ai-visibility"
        ],
        "summary": "Start Check",
        "operationId": "start_check_ai_visibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiVisibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiVisibilityCheck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/prospect": {
      "get": {
        "tags": [
          "prospect"
        ],
        "summary": "List Prospects",
        "operationId": "list_prospects_prospect_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProspectSheet"
                  },
                  "type": "array",
                  "title": "Response List Prospects Prospect Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "prospect"
        ],
        "summary": "Start Prospect",
        "operationId": "start_prospect_prospect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectSheet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/llms-txt": {
      "post": {
        "tags": [
          "llms-txt"
        ],
        "summary": "Build",
        "operationId": "build_llms_txt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmsTxtRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmsTxtResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/bulk": {
      "get": {
        "tags": [
          "bulk"
        ],
        "summary": "List Bulks",
        "operationId": "list_bulks_bulk_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BulkAudit"
                  },
                  "type": "array",
                  "title": "Response List Bulks Bulk Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "bulk"
        ],
        "summary": "Start Bulk",
        "operationId": "start_bulk_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkAudit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/sitemap-watcher": {
      "get": {
        "tags": [
          "sitemap-watcher"
        ],
        "summary": "List Watches",
        "operationId": "list_watches_sitemap_watcher_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SitemapWatch"
                  },
                  "type": "array",
                  "title": "Response List Watches Sitemap Watcher Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "sitemap-watcher"
        ],
        "summary": "Add Or Refresh",
        "operationId": "add_or_refresh_sitemap_watcher_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SitemapWatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitemapWatch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/perf-monitor": {
      "get": {
        "tags": [
          "perf-monitor"
        ],
        "summary": "List Monitors",
        "operationId": "list_monitors_perf_monitor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PerfMonitor"
                  },
                  "type": "array",
                  "title": "Response List Monitors Perf Monitor Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "perf-monitor"
        ],
        "summary": "Watch",
        "operationId": "watch_perf_monitor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfMonitorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfMonitor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/seo-tracker": {
      "get": {
        "tags": [
          "seo-tracker"
        ],
        "summary": "List Campaigns",
        "operationId": "list_campaigns_seo_tracker_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeoCampaign"
                  },
                  "type": "array",
                  "title": "Response List Campaigns Seo Tracker Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "seo-tracker"
        ],
        "summary": "Create",
        "operationId": "create_seo_tracker_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeoCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeoCampaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/scheduler": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get Status",
        "operationId": "get_status_scheduler_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Status Scheduler Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/auth/verify": {
      "get": {
        "summary": "Auth Verify",
        "description": "200 if creds OK, 401 otherwise. Used by the login screen / token checks.",
        "operationId": "auth_verify_auth_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Auth Verify Auth Verify Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Auth Token",
        "description": "Exchange the shared password for an opaque session token.\n\nBody: {\"password\": \"...\"}. On success returns {\"token\", \"expiresIn\"} so\nthe client can store the token instead of the raw password.",
        "operationId": "auth_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Auth Token Auth Token Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Auth Logout",
        "description": "Revoke a session token. Body: {\"token\": \"...\"}.",
        "operationId": "auth_logout_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Auth Logout Auth Logout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditJobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "globalScore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Globalscore"
          },
          "globalVerdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Globalverdict"
          },
          "criticalCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticalcount"
          },
          "warningCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warningcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "domain",
          "createdAt",
          "status"
        ],
        "title": "AuditJobSummary",
        "description": "Lightweight view of an audit — used in lists/sidebar."
      },
      "AuditRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "maxPages": {
            "type": "integer",
            "title": "Maxpages",
            "default": 300
          },
          "platform": {
            "type": "string",
            "enum": [
              "unknown",
              "custom",
              "webflow",
              "wordpress",
              "shopify",
              "wix",
              "squarespace",
              "bubble",
              "framer",
              "nextjs",
              "other"
            ],
            "title": "Platform",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AuditRequest"
      },
...
      "CompetitorBattle": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "targetUrl": {
            "type": "string",
            "title": "Targeturl"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/CompetitorSite"
            },
            "type": "array",
            "title": "Competitors"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitorReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "targetUrl",
          "competitors",
          "createdAt",
          "status"
        ],
        "title": "CompetitorBattle",
        "description": "A comparison between one target site and N competitors."
      },
      "CompetitorBattleRequest": {
        "properties": {
          "targetUrl": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Targeturl"
          },
          "competitors": {
            "items": {
              "type": "string",
              "maxLength": 2083,
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Competitors"
          }
        },
        "type": "object",
        "required": [
          "targetUrl",
          "competitors"
        ],
        "title": "CompetitorBattleRequest"
      },
...
      "ContentBrief": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "fr-FR"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "serpResults": {
            "items": {
              "$ref": "#/components/schemas/SerpResult"
            },
            "type": "array",
            "title": "Serpresults"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentBriefResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "query",
          "createdAt",
          "status"
        ],
        "title": "ContentBrief",
        "description": "A brief job — request + status + serp data + result."
      },
...
      "ContentBriefRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Query"
          },
          "locale": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Locale",
            "default": "fr-FR"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ContentBriefRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LlmsTxtRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "LlmsTxtRequest"
      },
...
      "ProspectRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ProspectRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}